Update Validate cho Add, Update, Delete

+++ excercise25_03/miniTest-JPA/src/main/java/vn/fis/traning/service/impl/AccountServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import lombok.Data;
 import vn.fis.traning.entity.Account;
 import vn.fis.traning.repo.AccountRepo;
 import vn.fis.traning.service.AccountService;
@@ -15,17 +16,47 @@ public class AccountServiceImpl implements AccountService{
 	
 	@Override
 	public Account addAccount(Account account) {
+		List<Account> list=accountRepo.findByAccountNumber(account.getAccountNumber());
+		if(list.size()!=0) {
+			return null;
+		}
+		if(account.getAccountNumber().length()!=12) {
+			return null;
+		}
+		if(account.getAccountName().length()<=5||account.getAccountName().length()>=100) {
+			return null;
+		}
+		account.setBalance(100000.0);
+		account.setStatus(1);
 		return accountRepo.save(account);
+		
 	}
 
 	@Override
 	public Account updateAccount(Account account) {
-		return accountRepo.save(account);
+		Account kq=accountRepo.findById(account.getAccountId()).orElse(null);
+//		List<Account> list=accountRepo.findByAccountNumber(kq.getAccountNumber());
+//		if(list.size()==0) {
+//			return null;
+//		}
+//		if(account.getAccountName().length()<=5||account.getAccountName().length()>=100) {
+//			return null;
+//		}
+//		if(account.getStatus()!=0&&account.getStatus()!=1&&account.getStatus()!=2) {
+//			return null;
+//		}
+		kq.setAccountName(account.getAccountName());
+		kq.setStatus(account.getStatus());
+		return accountRepo.save(kq);
 	}
 
 	@Override
-	public void deleteAccount(Long acconutId) {
-		accountRepo.deleteById(acconutId);
+	public String deleteAccount(Long acconutId) {
+		if(accountRepo.findById(acconutId)!=null) {
+			accountRepo.deleteById(acconutId);
+			return "Xóa thành công ^_^";
+		}
+		return "Tài khoản không hợp lệ!";
 	}
 
 	@Override
